# api.yaml
openapi: 3.0.0
info:
  title: Test API
  description: Testing API definition and code generation
  version: 1.0.0
components:
  parameters:
    limit:
        name: limit
        in: query
        description: Limits the number of returned results
        required: false
        schema:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
    offset:
        name: offset
        in: query
        description: Offset from which start returned results
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
  schemas:
    user:
      type: object
      description: A user record
      required:
        - uid
        - name_first
        - name_last
      properties:
        uid:
          type: string
          description: User ID
        name_first:
          type: string
          description: First name
        name_last:
          type: string
          description: Last name
        email:
          type: string
          format: email
          description: Email address
paths:
  /user/list:
    get:
      summary: Returns a list of users.
      description: List of user records.
      tags:
        - user
      parameters:
        - name: collab
          in: query
          description: Flag to return only collaborators
          required: false
          schema:
            type: boolean
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':    # status code
          description: A JSON array of user records
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/user'
  /user/view:
    get:
      summary: View a user record.
      description: View a user record.
      tags:
        - user
      parameters:
        - name: id
          in: query
          description: User ID of user record to view
          required: true
          schema:
            type: string
      responses:
        '200':    # status code
          description: A JSON user record
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/user'